requirements.txt
pdfplumber==0.9.0
python-docx==0.8.11
transformers==4.30.2
# torch==2.0.1
torch==2.6.0
jieba==0.42.1
tkinter==0.1.0
numpy==1.24.3
scikit-learn==1.3.0